wip: Split Session into model and view
The new `ClientManager` is a model responsible for managing clients.
A Client has three states: `Auth`, `Session`, and `LoggingOut`.

Auth in turn also has several states: `WaitPhoneNumber`, `WaitCode`,
etc.
Thus, login.rs is split up in several parts to represent these states.

The view just needs to display the correct widget for the current client
state.

The PR is ca. 50% done: Using the last logged in session should work.
The basic structure for the login is in place, but needs some more work.
Also, as there is now a single Bin to represent the client state, it
would be
nice to have a custom Bin widget that is able to show transitions on
`set_child`. Maybe it could use an internal Stack that removes the old
widget when the transition has finished.